+++ bot/states/__init__.py
@@ -1,0 +1,9 @@
+from bot.states.custom import CustomServerStates
+from bot.states.monitor import MonitorStates
+from bot.states.recharge import RechargeStates
+
+__all__ = [
+    'CustomServerStates',
+    'MonitorStates',
+    'RechargeStates',
+]